refactor(personopplysninger): egen fil for delte layoutkomponenter
affects: @navikt/familie-soknadsdialog-personopplysninger
